Fixed my part of the service layer as well as the reviews was an database issue.
