+++ backend/OfficeCalendar.Api/Controllers/EventParticipationsController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using OfficeCalendar.Api.Models;
 using OfficeCalendar.Api.Repositories;
+using OfficeCalendar.Api.Services;
 
 namespace OfficeCalendar.Api.Controllers
 {
@@ -9,68 +10,87 @@ namespace OfficeCalendar.Api.Controllers
     [Route("api/[controller]")]
     public class EventParticipationsController : ControllerBase
     {
-        private readonly GenericRepository<EventParticipation> _repository;
+        private readonly EventParticipationsService _service;
 
-        public EventParticipationsController(GenericRepository<EventParticipation> repository)
+        public EventParticipationsController(EventParticipationsService service)
         {
-            _repository = repository;
+            _service = service;
         }
 
         [HttpGet]
         public async Task<ActionResult<List<EventParticipation>>> GetAll()
         {
-            var participations = await _repository.GetAllAsync();
+            var participations = await _service.GetAllService();
             return Ok(participations);
         }
 
         [HttpGet("{id}")]
         public async Task<ActionResult<EventParticipation>> GetById(int id)
         {
-            var participation = await _repository.GetByIdAsync(id);
-            if (participation == null) return NotFound();
-            return Ok(participation);
+            try
+            {
+                var participation = await _service.GetByIdService(id);
+                return Ok(participation);
+            }
+            catch(KeyNotFoundException)
+            {
+                return NotFound();
+            }
         }
 
         [HttpGet("event/{eventId}")]
         public async Task<ActionResult<List<EventParticipation>>> GetByEvent(int eventId)
         {
-            var participations = await _repository.QueryAsync(
-                "SELECT * FROM event_participations WHERE event_id = @EventId",
-                new { EventId = eventId }
-            );
+            var participations = await _service.GetByEventService(eventId);
             return Ok(participations);
         }
 
         [HttpPost]
         public async Task<ActionResult<EventParticipation>> Create([FromBody] EventParticipation participation)
         {
-            participation.CreatedAt = DateTime.UtcNow;
-            var id = await _repository.InsertAsync(participation);
-            var created = await _repository.GetByIdAsync(id);
+            var (id, created) = await _service.CreateService(participation);
             return CreatedAtAction(nameof(GetById), new { id }, created);
         }
 
         [HttpPut("{id}")]
         public async Task<IActionResult> Update(int id, [FromBody] EventParticipation participation)
         {
-            var success = await _repository.UpdateAsync(id, participation);
-            if (!success) return NotFound();
-            return NoContent();
+            try
+            {
+                var success = await _service.UpdateService(id, participation);
+                return NoContent();
+            }
+            catch(KeyNotFoundException)
+            {
+                return NotFound();
+            }
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
-            var success = await _repository.DeleteAsync(id);
-            if (!success) return NotFound();
-            return NoContent();
+            try
+            {
+                var success = await _service.DeleteService(id);
+                return NoContent();
+            }
+            catch(KeyNotFoundException)
+            {
+                return NotFound();
+            }
         }
         [HttpDelete("{userId}/{eventId}")]
         public async Task<IActionResult> DeleteEventParticipationByUserIdAndEventId(int userId, int eventId)
         {
-            var success = await _repository.DeleteEventParticipationAsync(userId, eventId);
-            if (!success) return NotFound();
-            return NoContent();
-        }
+            try
+            {
+                var success = await _service.DeleteEventParticipationByUserIdAndEventIdAsyncService(userId, eventId);
+                return NoContent();
+            }
+            catch(KeyNotFoundException)
+            {
+                return NotFound();
+            }
+        }   
     }
 }
